Some initial work on improving add-in localization
The goal is to allow add-ins to be translated on Weblate alongside the main Pinta project, and then also ship these translations with the add-in. Using resource files rather than gettext works better, since the 'mautil pack' command for building the addin will automatically pull in the satellite assemblies for the resources. The string table localizer is an adhoc format that isn't understood by any other tools like Weblate

- Added missing translation strings for the effect properties.
- Switch to a string resource localizer for the reasons mentioned above.
- For now, the add-in header properties (description, etc) are left as-is in the XML file. It doesn't seem like there is a way to get these to be translated by the addin's localizer, so the solution is probably to generate a new .addin-xml automatially from the .resx files.